Fix handling of file switching for event meta data and unify tests
Make write / read tests reuse the same function

Significantly reduces code dupliation by having the write and
processEvent functions defined only once and re-used throughout all the
read and write tests.

Also increases test coverage in some places (e.g. MetaData have only
been partially covered until now) and reduces future maintenance costs.
